+++ fe/subsystems/gameState/gameStateMachine.cpp
@@ -72,7 +72,7 @@ void fe::gameStateMachine::push(baseGameState *newState, stateOptions options)
         m_endState->m_options = options;
 
         // we set the offset here so we can free the memory past the previous state, and behind the current
-        m_endState->m_offset = fe::memoryManager::get().getStackAllocater().getMarker();
+        m_endState->m_offset = static_cast<unsigned int>(fe::memoryManager::get().getStackAllocater().getMarker());
 
         m_update = false;
     }
