+++ fe/subsystems/graphic/renderObject/renderObject.hpp
@@ -55,7 +55,7 @@ namespace fe
                 char m_string[FE_MAX_STRING_RENDER_SIZE];
                 fe::fontData m_fontData;
                 float m_size[2];
-                float m_fontSize;
+                unsigned int m_fontSize;
                 unsigned int m_strSize;
                 unsigned int m_strPos;
